improve leaderboard indexer logic
