+++ indexers/leaderboard/src/mappings/mappingHandlers.ts
@@ -1,5 +1,7 @@
 import { SubstrateBlock } from "@subql/types";
 import * as db from "./db";
+import { EVENT_HANDLERS } from "./eventHandler";
+import { EXTRINSIC_HANDLERS } from "./extrinsicHandler";
 
 export async function handleBlock(_block: SubstrateBlock): Promise<void> {
   const {
@@ -7,51 +9,35 @@ export async function handleBlock(_block: SubstrateBlock): Promise<void> {
       header: { number },
       extrinsics,
     },
-    timestamp,
+    timestamp: blockTimestamp,
     events,
   } = _block;
   const height = BigInt(number.toString());
+  const timestamp = blockTimestamp ? blockTimestamp : new Date();
 
   let cache = db.initializeCache();
   let eventIndex = 0;
 
-  const [eventsByExtrinsic, finalizationEvents] = events.reduce<
-    [Record<number, typeof events>, typeof events]
-  >(
-    (acc, event) => {
-      if (event.phase.isApplyExtrinsic) {
-        const idx = event.phase.asApplyExtrinsic.toNumber();
-        if (!acc[0][idx]) acc[0][idx] = [];
-        acc[0][idx].push(event);
-      } else if (event.phase.isFinalization) {
-        acc[1].push(event);
+  const eventsByExtrinsic = new Map<number, typeof events>();
+  const finalizationEvents: typeof events = [];
+  for (const event of events) {
+    if (event.phase.isApplyExtrinsic) {
+      const idx = event.phase.asApplyExtrinsic.toNumber();
+      if (!eventsByExtrinsic.has(idx)) {
+        eventsByExtrinsic.set(idx, []);
       }
-      return acc;
-    },
-    [{}, []]
-  );
-
+      eventsByExtrinsic.get(idx)!.push(event);
+    } else if (event.phase.isFinalization) {
+      finalizationEvents.push(event);
+    }
+  }
   // Process extrinsics
   extrinsics.forEach((extrinsic, extrinsicIdx) => {
-    const extrinsicEvents = eventsByExtrinsic[extrinsicIdx] || [];
-    const { successEvent } = extrinsicEvents.reduce(
-      (
-        acc: {
-          successEvent?: (typeof events)[number];
-        },
-        event
-      ) => {
-        // Check for success event
-        if (
-          !acc.successEvent &&
-          event.event.section === "system" &&
-          event.event.method === "ExtrinsicSuccess"
-        ) {
-          acc.successEvent = event;
-        }
-        return acc;
-      },
-      {}
+    const extrinsicEvents = eventsByExtrinsic.get(extrinsicIdx) || [];
+    const successEvent = extrinsicEvents.findLast(
+      (event) =>
+        event.event.section === "system" &&
+        event.event.method === "ExtrinsicSuccess"
     );
     const successEventId = successEvent?.event.index.toString() || "";
     const extrinsicId = extrinsic ? height + "-" + extrinsicIdx.toString() : "";
@@ -80,275 +66,36 @@ export async function handleBlock(_block: SubstrateBlock): Promise<void> {
       );
 
       // Process specific extrinsic
-      switch (`${extrinsic.method.section}.${extrinsic.method.method}`) {
-        case "system.remark":
-        case "system.remarkWithEvent": {
-          const eventRemarked = extrinsicEvents.find(
-            (e) => e.event.section === "system" && e.event.method === "Remarked"
-          );
-          const eventRemarkedId = eventRemarked
-            ? height + "-" + eventRemarked.event.index.toString()
-            : "";
-          cache.accountRemarkCountHistory.push(
-            db.createAccountRemarkCount(
-              extrinsicSigner,
-              BigInt(1),
-              height,
-              extrinsicId,
-              eventRemarkedId,
-              timestamp
-            )
-          );
-          break;
-        }
-      }
+      const extrinsicKey =
+        extrinsic.method.section + "." + extrinsic.method.method;
+      const handler = EXTRINSIC_HANDLERS[extrinsicKey];
+      if (handler)
+        handler({
+          extrinsicEvents,
+          cache,
+          height,
+          timestamp,
+          extrinsicId,
+          extrinsicSigner,
+        });
 
       // Process extrinsic events
       extrinsicEvents.forEach((event, eventIdx) => {
         const eventId = height + "-" + eventIndex;
 
         // Process specific events
-        switch (`${event.event.section}.${event.event.method}`) {
-          case "balances.Transfer": {
-            const from = event.event.data[0].toString();
-            const to = event.event.data[1].toString();
-            const amount = BigInt(event.event.data[2].toString());
-
-            cache.accountTransferSenderTotalCountHistory.push(
-              db.createAccountTransferSenderTotalCount(
-                from,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.accountTransferSenderTotalValueHistory.push(
-              db.createAccountTransferSenderTotalValue(
-                from,
-                amount,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-
-            cache.accountTransferReceiverTotalCountHistory.push(
-              db.createAccountTransferReceiverTotalCount(
-                to,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.accountTransferReceiverTotalValueHistory.push(
-              db.createAccountTransferReceiverTotalValue(
-                to,
-                amount,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          case "transactionPayment.TransactionFeePaid": {
-            const who = event.event.data[0].toString();
-            const actualFee = BigInt(event.event.data[1].toString());
-            const tip = BigInt(event.event.data[2].toString());
-
-            cache.accountTransactionFeePaidTotalValueHistory.push(
-              db.createAccountTransactionFeePaidTotalValue(
-                who,
-                actualFee + tip,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          case "domains.OperatorRewarded": {
-            const operatorId = event.event.data[1].toString();
-            const reward = BigInt(event.event.data[1].toString());
-            if (reward === BigInt(0)) break;
-
-            cache.operatorTotalRewardsCollectedHistory.push(
-              db.createOperatorTotalRewardsCollected(
-                operatorId,
-                reward,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          case "domains.OperatorTaxCollected": {
-            const operatorId = event.event.data[0].toString();
-            const tax = BigInt(event.event.data[1].toString());
-
-            cache.operatorTotalTaxCollectedHistory.push(
-              db.createOperatorTotalTaxCollected(
-                operatorId,
-                tax,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-            break;
-          }
-          case "domains.BundleStored": {
-            const bundleAuthor = event.event.data[0].toString();
-
-            cache.operatorBundleTotalCountHistory.push(
-              db.createOperatorBundleTotalCount(
-                bundleAuthor,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          case "domains.OperatorRegistered": {
-            const domainId = event.event.data[0].toString();
-            const operatorId = event.event.data[1].toString();
-            const amount = BigInt(extrinsic.data[0].toString());
-            const nominatorId = extrinsic.signer.toString();
-
-            cache.operatorDepositsTotalCountHistory.push(
-              db.createOperatorDepositsTotalCount(
-                operatorId,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.operatorDepositsTotalValueHistory.push(
-              db.createOperatorDepositsTotalValue(
-                operatorId,
-                amount,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-
-            cache.nominatorDepositsTotalCountHistory.push(
-              db.createNominatorDepositsTotalCount(
-                nominatorId,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.nominatorDepositsTotalValueHistory.push(
-              db.createNominatorDepositsTotalValue(
-                nominatorId,
-                amount,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          case "domains.OperatorNominated": {
-            const operatorId = event.event.data[0].toString();
-            const nominatorId = event.event.data[1].toString();
-            const amount = BigInt(event.event.data[2].toString());
-
-            cache.operatorDepositsTotalCountHistory.push(
-              db.createOperatorDepositsTotalCount(
-                operatorId,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.operatorDepositsTotalValueHistory.push(
-              db.createOperatorDepositsTotalValue(
-                operatorId,
-                amount,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-
-            cache.nominatorDepositsTotalCountHistory.push(
-              db.createNominatorDepositsTotalCount(
-                nominatorId,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.nominatorDepositsTotalValueHistory.push(
-              db.createNominatorDepositsTotalValue(
-                nominatorId,
-                amount,
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          case "domains.WithdrewStake": {
-            const operatorId = event.event.data[0].toString();
-            const nominatorId = event.event.data[1].toString();
-
-            cache.operatorWithdrawalsTotalCountHistory.push(
-              db.createOperatorWithdrawalsTotalCount(
-                operatorId,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            cache.nominatorWithdrawalsTotalCountHistory.push(
-              db.createNominatorWithdrawalsTotalCount(
-                nominatorId,
-                BigInt(1),
-                height,
-                extrinsicId,
-                eventId,
-                timestamp
-              )
-            );
-            break;
-          }
-          default:
-            break;
-        }
+        const eventKey = event.event.section + "." + event.event.method;
+        const handler = EVENT_HANDLERS[eventKey];
+        if (handler)
+          handler({
+            event,
+            extrinsic,
+            cache,
+            height,
+            timestamp,
+            extrinsicId,
+            eventId,
+          });
 
         // Increment event index
         eventIndex++;
@@ -374,106 +121,17 @@ export async function handleBlock(_block: SubstrateBlock): Promise<void> {
     const eventId = height + "-" + eventIndex;
 
     // Process specific events
-    switch (`${event.event.section}.${event.event.method}`) {
-      case "rewards.VoteReward": {
-        const voter = event.event.data[0].toString();
-        const reward = BigInt(event.event.data[1].toString());
-
-        cache.farmerVoteTotalCountHistory.push(
-          db.createFarmerVoteTotalCount(
-            voter,
-            BigInt(1),
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-        cache.farmerVoteTotalValueHistory.push(
-          db.createFarmerVoteTotalValue(
-            voter,
-            reward,
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-
-        cache.farmerVoteAndBlockTotalCountHistory.push(
-          db.createFarmerVoteAndBlockTotalCount(
-            voter,
-            BigInt(1),
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-        cache.farmerVoteAndBlockTotalValueHistory.push(
-          db.createFarmerVoteAndBlockTotalValue(
-            voter,
-            reward,
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-
-        break;
-      }
-      case "rewards.BlockReward": {
-        const blockAuthor = event.event.data[0].toString();
-        const reward = BigInt(event.event.data[1].toString());
-
-        cache.farmerBlockTotalCountHistory.push(
-          db.createFarmerBlockTotalCount(
-            blockAuthor,
-            BigInt(1),
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-        cache.farmerBlockTotalValueHistory.push(
-          db.createFarmerBlockTotalValue(
-            blockAuthor,
-            reward,
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-
-        cache.farmerVoteAndBlockTotalCountHistory.push(
-          db.createFarmerVoteAndBlockTotalCount(
-            blockAuthor,
-            BigInt(1),
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-        cache.farmerVoteAndBlockTotalValueHistory.push(
-          db.createFarmerVoteAndBlockTotalValue(
-            blockAuthor,
-            reward,
-            height,
-            extrinsicId,
-            eventId,
-            timestamp
-          )
-        );
-
-        break;
-      }
-      default:
-        break;
-    }
+    const eventKey = event.event.section + "." + event.event.method;
+    const handler = EVENT_HANDLERS[eventKey];
+    if (handler)
+      handler({
+        event,
+        cache,
+        height,
+        timestamp,
+        extrinsicId,
+        eventId,
+      });
 
     // Increment event index
     eventIndex++;
