+++ indexers/leaderboard/tsconfig.json
@@ -10,6 +10,7 @@
     "outDir": "dist",
     "rootDir": "src",
     "target": "es2017",
+    "lib": ["ES2023"],
     "strict": true
   },
   "include": [
